Added UIDrawerRowLayout, completed scene load/save, editor components finished
Other things added: reset for EntityManager and AssetDatabaseReference
Reworked editor components such that it reflects the state of the data inside the sandboxes. It handled the internal events on multiple threads and returns to the user a list with events that cannot be handled internally like removing a component and clashing IDs such that it asks the user to change the code or to choose what happens.