Asset editor integration
The loading of assets should mostly be done now. All the fixups and remaps for database references and link component splats should be good to go (a little more testing can be done).
Misc and material assets settings UI needs to be finished
GLTF now can coallesce the mesh on the CPU and upload the data directly to the GPU without any syncronization or locks of the immediate context in multithreaded contexts.