Automatic asset reloading finished
Automatic asset reloading has been finished.
To be done are still the material settings inspector + the misc one.
Then the assets are pretty much done.
The next step is Graphics module integration
Another thing to be noted is that we need to create the UI graphics
on the dedicated GPU as well since using the dedicated GPU for rendering of the sandboxes will mean that we have to read the texture to RAM and this create a lot of wasted bandwidth. The overhead of the UI should be small enough not to cause any major problems (only on battery that is a valid concern).