Improved Shader Reflection and working towards material integration
Shader Reflection can now create reflection types for constant buffers.
Utility string functions + a copy with memset.
Moved the Copy function for SerializeCustomType into ReflectionCustomType.
Modules now correctly take into account their dependencies such that they can still be recompiled. (before it would cause an error when recompiling a .dll which was referenced somewhere else because the file would still be locked). Also recompilation will trigger the dependencies to be compiled first, then the main module.
Materials now are in bare mode functional. There are still missing callbacks for changed values and the way unloading/loading of assets is done when the material is not referenced.