The material asset is finished, small bug fixes for assets are still required
The material is now finished along all its problems (dealing with dependencies).
Added functions for hash table deep copy and deallocation based on templates.
Small fix to the InspectorEntity to destroy the UIInstance when removing a component.
Quality of life changes to the asset database for dealing with dependencies. Also fixed a bug when a material is unloaded but its dependencies are not.
Added a function to convert a material from a database to another database.
Reworked material file serialization/deserialization to support constant buffers.
FIxed a bug for the resource manager when loading/unloading user materials and changed the initial increment from 100 to 1 for resources that are referenced counted. (Why was it 100 before anyway?).
Modified the Graphics functions that create shaders to be able to fail by returning nullptr instead of crashing. Also removed the Shader target 5.1 because that is used by D3D12.
Completely revamped the ShaderReflection to be able to support generating ReflectionTypes for constant buffers and to be able to ignore comments and use macros for conditional inclusion.
Small fixes inside UIDrawer for elements.
UIReflection now supports groupings with name overrides such that it makes easier having multiple fields treated as the same entity (useful for example for matrices).
Inside Function added functions for removal of single line comments, multi line comments and conditional inclusion of text like a C preprocessor.
DeserializeFieldTable now allows generating ReflectionTypes that can be commited into a reflection manager.
Custom serialization types now can have boolean switches that they can access easily without creating compound structures.
Restructured the Editor Asset registering/unregistering functions to be defined in the AssetManagement and allow registering/unregistering of assets that are not tied to a specific sandbox.
Fixed a multithreading bug in the EditorComponents created in the FinalizeEvent when creating UITypes. Because it would run on a background thread it could happen that the main thread is also running UIReflection code and create a race condition. Now FinalizeEvent will push the creation into an EditorEvent.
AssetOverrides now support retargetting of the asset database (they don't go now directly to the main database of the editor).