Rendering continuation
DebugDrawer added to a World.
Inlined suitable functions
Started working on GPUProfiling.
DebugDrawer axes rescaled such that when size is set to 1.0f it will be the same size as a sphere with a radius of 1.0f.
EditorSandbox preinitialization now waits for the graphics creation before creating the sandbox world.
Small bug for material inspector UI that resulted in the reflection manager being allocated from a temporary allocator and not recovered.
Scene now corrects the pacing when using right click (rotation) such that it provides smooth feedback