Continuing module rendering
Replaced AddSafe with AddAssert for CapacityStream
ResourceManager now checks that shaders exists for unloading user material
Small Reflection bug that caused stream_type_byte_size to be omitted for user defined types that were contained by streams.
New serialization field that is written in the deserialization table, pointer offset, that helps reconstruct 1:1 the reflection_type (previously the pointer was presumed to be contiguous from the last field but for certain types - like those for a constant buffer from a shader - it wasn't the case)
 Re-rendering of the sandbox now is mostly performed - on component change, on asset change, on asset metadata change.
Fixed bugs regarding Asset reloading - it would fail sometimes if previously the asset wasn't built or afterwards.
Viewport rendering now can be enabled and disabled (used by the UI windows)
Some synchronization issues are still present when rendering the viewport - changed the priorities of the rendering threads to very low to help them being identified from the rest in the debugging window
Restoring a hub project now correctly identifies the name from the path