More asset work
Material options can now be constructed from reflected shader
The asset database will update its assets that have metadata dependencies on target files - like materials that can have reflected properties that have changed on disk but they are different in the serialized scene version.
Fixed a bug where multiple threads would be accessing the RuntimeGraphics immediate context - when resizing a viewport and loading resources - the resizing wouldn't check for that.
Now materials updates reflect directly in the viewport rather when the input field is released (for sliders, number input and so on).