Worked towards Visualize Texture window to help diagnose GPU texture problems
Base VisualizeTexture window is done. Integration with the Editor is almost complete. Next there needs to be implemented controls for RGBA fields and a way to let you change the selected texture based on a list given.
Improvements were made to the GRAPHICS_FORMAT functions to support this.
New compute helper shaders were added to perform this visualization