DebugInfo added to calls for more info
Most containers receive now DebugInfo that is passed towards the allocation calls in order to have better diagnostics
FIxed a design flaw in the EditrState event queue. Previously, it was a resizable queue with a lock to protect it. But there were problems in how the locking was done leading to race conditions. Now the queue is a resizable thread safe queue which protects itself accordingly.