Continued work towards Focus object
More QoL improvements to containers and VCLExtensions
Fixed DeserializeEntityManager which had 2 paths: one with uintptr_t and another one with ECS_FIILE_HANDLE which were separate and very annoying to maintain since having to make modifications to both of them and most of them I would forget about the 2nd variant. Now they have been unified.
Modules now have an extra function named RegisterExtraInformation. This is just a flattened hash table (just an array of pairs of key and value) that can be used to inform the Editor/Runtime about miscellaneous stuff. The only consideration is that everything needs to be turned into a string and parsed again, but for small information it is perfect since it is so lightweight.
Now we can have link components in the ECSEngine Components.h
The camera Focus function is finished but untested