Fixed UI window Undo queue, started working on link components delta
The UI revert/undo queue now has an owning pointer associated with each command. Before executing it, we check to see if the pointer is still a valid memory resource. If it is, then we allow the execution, otherwise we discard it.
Refactored some code that could use newer constructs to perform the same task