Entity Manager global component, ECS Camera component, Editor Sandbox Modules additions
The entity manager now can hold global components (one per each entity manager instance that are can be easily retrieved) that are serialized.
The entity manager serializer has been updated to accomodate this feature
The reflection system and the link components have been updated to accomodate global ECS components
New Sandbox module functions to help with detection of changes while a sandbox is running in order to prevent state from previous DLLs being kept in the task manager and task scheduler (this is still in progress)
Now runtime resources can be queried to determine if it runs in SCENE or GAME context inside the editor
The runtime can now handle the case where assets are changed during runtime that would cause scene assets to be removed and when the state would be restored these would be invalid. This case is handled by doubling the reference count of the assets for the sandbox that is about to be run such that these assets are pinned
The entity inspector can now accomodate dynamically  removed components without being notified