RunWorld snapshots have been implemented
The snapshots taken at each simulation step have been implemented. They are checked before the runtime proceeds with the frame such that we don't continue if there are modules missing, reloaded modules, compiling modules or settings which have been changed for modules.
Added a flags field to the sandbox for miscellaneous signaling. At the moment, it will be used to signal to the UI that the runtime is waiting for the compilation of the modules.