Continued integrating global components, module compilation and fixing editor component events
Integrated ECS_COMPONENT_TYPE into APIs to make it easier to work with all types of components
Fixed errors in the module build phases. We cannot wait for modules to finish compiling since they might push events, so we cannot stall the main thread. This is fine tho - it is not good practice to stall the entire editor to finish compilation since it takes a lot of time. Fixed EDITOR_FINISH_BUILD_COMMAND_STATUS being given without the atomic specifier - now they are explicit atomic to indicate that they are modified by multiple threads
GLTFLoader now correctly resizes the submesh bounds when a scale factor is specified (need to check if all functions do this tho)
 Fixed asset reloading that updated components only in the active entity manager. The update needs to be performed for the scene entity manager always and for the runtime manager if it is in running or paused state.
Editor component events had to be improved a lot. New PROMOTED_TO_COMPONENT and COMPONENT_DEMOTED to handle the case when it has the tag but no IsShared function. The events' names are always allocated now to make easier the mental model since the IS_REMOVED event could be generated when the type would be actually removed, and this would result in the name being dangling. Also, made the link types accomodate these changed - needed to perform some tweaks to make this work