Sandbox asset management
Added new SoA functions to help work with SoA streams.
Fixed compression bugs: the texture when using GPU compression would not get put into the graphics bookkeeping. BC1, BC3 and BC7 missing srgb handling. Added a new check when compressing to make that the given data matches the compression (for example, when using BC6 it makes sure that there are enough bytes to be read, otherwise it can result in a segfault)
Fixed some bugs related to material loading both in the resource manager and asset loading. There is still to be figured out what to do when a material reload is happening and a failure is encountered for one of the reloads: should we unload the previous material or not? (currently unloading causes crashes since the old resource might be discarded)
Implemented sandbox asset handle snapshot that gets all the reference counts for the database assets and determines the delta between the new reference counts from entities at each frame and increments or decrements the count appropriately. Accommodations has to be made to the delete entity/copy entity/remove component/reset component functions and entity inspector asset callback to not handle the reference counts during a runtime for a sandbox.