Exporting material textures from embedded .glb
We can export the textures that are embedded in a .glb.
Now the engine can write .jpg files.
Kept adding more ECS_INLINE (where it is appropriate) and slowly adding static in .cpp to indicate internal functions
Changed names for some functions to better reflect their action
EntityManager now correctly copied the global components in the CopyOther function
Fixed inconsistencies in the asset path determining which caused assets metadatas to be missed
Split GraphicsHelpers.h into a new TextureOperations.h header to have better modularity
Fixed a bug in AssetLoading which caused an infinite wait on the semaphore when the thread waiting on the semaphore had tasks to be performed. Now the thread exhausts its thread queue before sleeping on the semaphore
Improved PBR material detection out of GLTF
Added SIMD Math types to ReflectionManager blittable exceptions to be easily skipped